cleaned up test directory
