+++ Compiler/test/klein/test_klein_program_1.kln
