+++ Compiler/test/klein/test_klein_program_multiple_function_calls.kln
